% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasummary2tex.R
\name{datasummary2tex}
\alias{datasummary2tex}
\title{Convert datasummary to LaTeX}
\usage{
datasummary2tex(
  x,
  var_title = "Variable",
  placement = "h!",
  caption,
  label,
  size = "normalsize",
  tablewidth = 0.98,
  note,
  file
)
}
\arguments{
\item{x}{Object of class \code{tinytable} created from \link[modelsummary:datasummary]{modelsummary::datasummary}.}

\item{var_title}{Character scalar, column title for variables.}

\item{placement}{Character scalar, placement option for LaTeX table, default is "h!".}

\item{caption}{Table caption. Default will leave it empty.}

\item{label}{Label. Default will leave it empty.}

\item{size}{Font size.}

\item{tablewidth}{Width of \code{tabularx} table (as percentage of \verb{\\textwidth}).}

\item{note}{Table note, default is empty.}

\item{file}{Path of file to be written. Default will print to console.}
}
\value{
A tex table that can be printed or saved to file.
}
\description{
Converts a \link[modelsummary:datasummary]{modelsummary::datasummary} object of class \code{tinytable} to LaTex.
While \code{datasummary} produces tex output on its own, the layout
options are restrictive. \code{datasummary2tex} allows for flexible layout options,
including combining several \code{datasummary} objects into one.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{library(modelsummary)

# Convert `datasummary` object to LaTeX
mod <- datasummary(
 hp + mpg ~
 Mean + SD + Min + Max + N,
 data = mtcars)

datasummary2tex(mod)
}\if{html}{\out{</div>}}
}

