% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_etable.R
\name{define_fitstats}
\alias{define_fitstats}
\title{Define fit statistics for etable}
\usage{
define_fitstats()
}
\value{
Returns invisibly.
}
\description{
Defines a set of custom fit statistics for use in \link[fixest:etable]{fixest::etable}, such as
the outcome mean, number of fixed effects groups, first stage IV stats.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{# library(fixest)

est1 = feols(Ozone ~ i(Month) / Wind + Temp, data = airquality)
est2 = feols(Ozone ~ i(Month, Wind) + Temp | Month, data = airquality)

define_fitstats()
etable(est1, est2,
       fitstat = c("n", "ar2", "y_mean"))
}\if{html}{\out{</div>}}
}

