% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coefs.R
\name{get_coefs}
\alias{get_coefs}
\title{Get fixest coefficients}
\usage{
get_coefs(
  model,
  levels = c(0.95, 0.9),
  keep,
  drop_constant = TRUE,
  append = FALSE,
  append_x
)
}
\arguments{
\item{model}{Model of class \code{fixest}.}

\item{levels}{Numeric vector of confidence levels. Must be strictly between 0 and 1.
For example \code{c(0.90, 0.95)}.}

\item{keep}{Character vector of coefficients to keep}

\item{drop_constant}{Logical, whether to drop the intercept.}

\item{append}{Logical, whether to append the output to \code{append_x}. If \code{TRUE} \code{append_x}
must be supplied.}

\item{append_x}{Dataframe to append the coefficient estimates (must be suppled if
\code{append} is \code{TRUE}).}
}
\value{
A dataframe of coefficient estimates and confidence intervals.
}
\description{
Get fixest coefficients
}
