% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsensemakr.R
\name{ggsensemakr}
\alias{ggsensemakr}
\title{Produce contour plots from sensitivity models}
\usage{
ggsensemakr(
  mod.sens,
  plot_type = c("estimate", "t-value", "extreme"),
  n_levels = 8,
  levels_range = c("range", "iqr"),
  label_hjust = 0.02,
  label_vjust = 0
)
}
\arguments{
\item{mod.sens}{Object of class \code{sensemakr}.}

\item{plot_type}{Type of sensitivity plot. It can be either "estimate" (default)
for contour plots with sensitivity of point estimates; "t-value" for contour plots
with sensitivity of t-values; or "extreme" for extreme scenarios plots.}

\item{n_levels}{The approximate number of contour lines to draw.}

\item{levels_range}{Whether the drawing of contour lines should use the full range of values
or only exclude the top- and bottom percentiles. This is mostly a convenience option to
to adjust the number of contour lines shown.}

\item{label_hjust}{Horizontal adjustment of labels.}

\item{label_vjust}{Vertical adjustment of labels.}
}
\value{
ggplot of contour lines.
}
\description{
Plots contour (and related) plots from sensivity analyses conducted
using \link[sensemakr:sensemakr]{sensemakr::sensemakr}. Elements of the plot can be styled subsequently
using standard \link[ggplot2:theme]{ggplot2::theme} elements.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{library(sensemakr)
library(dplyr)
library(stringr)
library(ggplot2)

mod <- lm(
  peacefactor ~ directlyharmed  + village +  female +
    age + farmer_dar + herder_dar + pastvoted + hhsize_darfur,
  data = darfur)

mod.sens <- sensemakr(
  model = mod,
  treatment = "directlyharmed",
  benchmark_covariates = "female",
  kd = 1:3)

# Sensivitity plot of estimates
ggsensemakr(mod.sens)

# Sensivitity plot of t-values
ggsensemakr(mod.sens, plot_type = "t-value")

# Sensivitity plot of extreme confounding
ggsensemakr(mod.sens, plot_type = "extreme")
}\if{html}{\out{</div>}}
}

