% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{my_maptheme}
\alias{my_maptheme}
\title{Custom ggplot2 map theme}
\usage{
my_maptheme(map_grid = FALSE, map_grid_labels = FALSE, ...)
}
\arguments{
\item{map_grid}{Logical, whether to show grid of longitude and latitude}

\item{map_grid_labels}{Logical, whether to show axis labels of longitude/latitude}

\item{...}{Additional parameters passed on to \link{my_theme}}
}
\value{
Custom ggplot2 map theme
}
\description{
Custom ggplot theme for maps.
}
