% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{my_theme}
\alias{my_theme}
\title{Custom ggplot2 theme}
\usage{
my_theme(
  font_family = "Arial Narrow",
  font_size = 16,
  rel_small = 14/16,
  rel_tiny = 12/16,
  rel_large = 18/16,
  grid_major_color = "grey90",
  grid_minor_color = "grey95",
  grid_major_linewidth = 0.4,
  grid_minor_linewidth = 0.3,
  axis_title_hjust = 0.97,
  axis_title_size = 18,
  axis_text_size = 16
)
}
\arguments{
\item{font_family}{Font family (requires \link[Cairo:Cairo]{Cairo::Cairo} package).}

\item{font_size}{Standard font size}

\item{rel_small}{Relative size of small font}

\item{rel_tiny}{Relative size of tiny font}

\item{rel_large}{Relative size of large font}

\item{grid_major_color}{Color of major grid lines}

\item{grid_minor_color}{Color of minor grid lines}

\item{grid_major_linewidth}{Width of major grid lines}

\item{grid_minor_linewidth}{Width of minor grid lines}

\item{axis_title_hjust}{Horizontal adjustment of axis titles}

\item{axis_title_size}{Font size of axis titles}

\item{axis_text_size}{Font size of axis test}
}
\value{
ggplot2 theme
}
\description{
Custom these for ggplot, building on \code{\link[cowplot:theme_minimal_grid]{cowplot::theme_minimal_grid()}}.
}
