% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_etable.R
\name{set_table_layout}
\alias{set_table_layout}
\title{Set etable postprocessing function}
\usage{
set_table_layout(
  x,
  size = "small",
  note,
  clustervar,
  subsample,
  tabularx = FALSE,
  tablewidth = 0.98,
  robust = F,
  conley = F,
  conley_cut,
  tabularY = FALSE,
  sidewaystable = FALSE,
  linespacing = 1,
  replace.name,
  conley.se.row = F,
  multipanel = F,
  panel_no = 1,
  panel_header = ""
)
}
\arguments{
\item{x}{\code{etable} character vector.}

\item{size}{Relative font size (e.g., normalsize, small, footnotesize).}

\item{note}{Table note.}

\item{clustervar}{Character, name of variable used to cluster standard errors.
This can be any character string, not necessarily the column name (which is specified
as part of the \code{fixest} formula.)}

\item{subsample}{??}

\item{tabularx}{Logical, whether to use \code{tabularx}.}

\item{tablewidth}{Relative width of table (to \verb{\\textwidth}).}

\item{robust}{Logical, if robust standard errors are used.}

\item{conley}{Logical, if Conley HAC standard errors are used.}

\item{conley_cut}{Integer, the distance cut-off used for Conley HAC standard errors.}

\item{tabularY}{Logical, whether to use \code{tabularY}.}

\item{sidewaystable}{Logical, whether to use \code{sidewaystable}.}

\item{linespacing}{Line space (in \code{em} units).}

\item{replace.name}{Vector of length 2 used to replace a certain name within the table. Should
be in the form \code{c("old_name", "new_name")}.}

\item{conley.se.row}{Whether a row of SE specifications should be used, which will change the
display of Conley HAC standard errors (and leave other columns with, e.g., regular SEs as is).}

\item{multipanel}{Logical, if table should be multipanel.}

\item{panel_no}{Integer, the number of panel in the multipanel table.}

\item{panel_header}{Character, header of the panel.}
}
\value{
Returns invisibly
}
\description{
Postprocessing options for creating custom LaTeX output from fixest models.
The arguments should be supplied in the regular \link[fixest:etable]{fixest::etable} with the
inclusion of \code{postprocess.tex = set_table_layout}.
}
\section{Examples}{


\if{html}{\out{<div class="sourceCode R">}}\preformatted{# library(fixest)

est1 = feols(Ozone ~ i(Month) / Wind + Temp, data = airquality)
est2 = feols(Ozone ~ i(Month, Wind) + Temp | Month, data = airquality)

etable(est1, est2, tex = T,
       postprocess.tex = set_table_layout,
       style.tex = table_style())
}\if{html}{\out{</div>}}
}

