% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_etable.R
\name{table_style}
\alias{table_style}
\title{Set table style for etable}
\usage{
table_style()
}
\value{
Returns invisibly
}
\description{
Set table style for use in \link[fixest:etable]{fixest::etable}. Table style should be supplied
in \link[fixest:etable]{fixest::etable} with the inclusion of \code{style.tex = table_style}.
}
